pequena correção

+++ src/Exercicios/AreaQuadrado.java
@@ -1,7 +1,7 @@
 package Exercicios;
 
 public class AreaQuadrado {
-    public void areaQuadrado(int lado){
+    public void areaQuadrado(double lado){
         double area = Math.pow(lado,2);
         System.out.printf("Área do quadrado: %.0f", area);
     }
